Type: chore(ut-desktop gridmanager canvasgridview ) 调整一下格式
Description:

RootCause:Null

Solution:
Former-commit-id: 1de27515e846722d0885742799317c6e32ad5508

+++ dde-desktop/tests/src/dde-wallpaper-chooser/ut-waititem-test.cpp
@@ -5,8 +5,8 @@
 #include <QLabel>
 #define private public
 
-#include "../../dde-wallpaper-chooser/waititem.h"
-#include "../../dde-wallpaper-chooser/waititem.cpp"
+#include "../dde-wallpaper-chooser/waititem.h"
+#include "../dde-wallpaper-chooser/waititem.cpp"
 
 using namespace testing;
 
@@ -44,7 +44,7 @@ TEST_F(WaitItemTest, initSize)
         int temp = sumwidth - (m_testitem->width() - static_cast<int>(m_testitem->m_movedistance));
         m_testitem->m_movedistance -= temp;
     }
-    int temp = m_testitem->m_movedistance;
+    int temp = static_cast<int>(m_testitem->m_movedistance);
 
     EXPECT_EQ(m_testitem->size(), QSize(400, 400));
     EXPECT_EQ(temp, static_cast<int>(m_testitem->m_movedistance));
